+++ superwall/src/main/java/com/superwall/sdk/paywall/presentation/PaywallInfo.kt
@@ -9,6 +9,7 @@ import com.superwall.sdk.dependencies.TriggerSessionManagerFactory
 import com.superwall.sdk.misc.camelCaseToSnakeCase
 import com.superwall.sdk.models.config.FeatureGatingBehavior
 import com.superwall.sdk.models.events.EventData
+import com.superwall.sdk.models.paywall.LocalNotification
 import com.superwall.sdk.models.product.Product
 import com.superwall.sdk.models.serialization.URLSerializer
 import com.superwall.sdk.models.triggers.Experiment
@@ -50,6 +51,7 @@ data class PaywallInfo(
     val isFreeTrialAvailable: Boolean,
     val featureGatingBehavior: FeatureGatingBehavior,
     val closeReason: PaywallCloseReason,
+    val localNotifications: List<LocalNotification>,
     val computedPropertyRequests: List<ComputedPropertyRequest>,
     val surveys: List<Survey>,
     val factory: TriggerSessionManagerFactory
@@ -77,6 +79,7 @@ data class PaywallInfo(
         presentationSourceType: String? = null,
         factory: TriggerSessionManagerFactory,
         featureGatingBehavior: FeatureGatingBehavior = FeatureGatingBehavior.NonGated,
+        localNotifications: List<LocalNotification>,
         computedPropertyRequests: List<ComputedPropertyRequest>,
         closeReason: PaywallCloseReason,
         surveys: List<Survey>
@@ -122,6 +125,7 @@ data class PaywallInfo(
             }
         },
         factory = factory,
+        localNotifications = localNotifications,
         computedPropertyRequests = computedPropertyRequests,
         closeReason = closeReason,
         surveys = surveys