+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The changelog for `Superwall`. Also see the [releases](https://github.com/superw
 
 ### Enhancements
 
+- SW-2214: Adds ability to use free trial notifications with a paywall.
 - SW-2640: Adds `computedPropertyRequests` to `PaywallInfo`.
 - SW-2641: Makes `closeReason` in `PaywallInfo` non-optional.